Expand `globalQueue` to handle state updates, too.
